+++ src/Containers/CMakeLists.txt
@@ -20,4 +20,5 @@ if(BUILD_UNIT_TESTS)
   add_subdirectory(OhmmsPETE/tests)
   add_subdirectory(OhmmsSoA/tests)
   add_subdirectory(MinimalContainers/tests)
+  add_subdirectory(Pools/tests)
 endif()
