+++ src/Containers/Pools/tests/CMakeLists.txt
@@ -11,24 +11,10 @@
 
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${QMCPACK_UNIT_TEST_DIR})
 
-set(SRC_DIR utilities)
-set(UTEST_EXE test_${SRC_DIR})
-set(UTEST_NAME deterministic-unit_test_${SRC_DIR})
+set(UTEST_EXE test_containers_pools)
+set(UTEST_NAME deterministic-unit_test_containers_pools)
 
-add_executable(
-  ${UTEST_EXE}
-  test_rng.cpp
-  test_parser.cpp
-  test_timer.cpp
-  test_runtime_manager.cpp
-  test_prime_set.cpp
-  test_partition.cpp
-  test_ResourceCollection.cpp
-  test_infostream.cpp
-  test_output_manager.cpp
-  test_StdRandom.cpp)
-target_link_libraries(${UTEST_EXE} catch_main qmcutil)
+add_executable(${UTEST_EXE} test_pooled_memory.cpp)
+target_link_libraries(${UTEST_EXE} catch_main containers)
 
 add_unit_test(${UTEST_NAME} 1 1 "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
-
-subdirs(for_testing)
